pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proyecto-final-deep"
version = "0.1.0"
description = "Forecasting meteorológico horario INMET con Deep Learning"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Juan Andrés Ramos" }]

dependencies = [
    # core
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "scipy>=1.13",
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "joblib>=1.3",
    # deep learning
    "torch>=2.2",
    # forecasting / stats
    "scikit-posthocs>=0.9",
    # tracking (opcional)
    "mlflow>=2.12",
    # jupyter
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "jupyter-book>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
filterwarnings = ["ignore::DeprecationWarning"]
